Moving setting requestId to SystemTraceHandler
